requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.68.2
pillow==11.1.0
opencv-python-headless==4.11.0.86
numpy==2.2.4
python-multipart==0.0.20
